Conflicts:
	.gitignore
	README.rst
	apps/activity/feeds.py
	apps/drumbeat/views.py
	apps/drumbeatmail/urls.py
	apps/projects/forms.py
	apps/projects/views.py
	apps/users/tasks.py
	apps/users/views.py
	media/css/batucada.css
	media/js/script.js
	settings.py
	templates/base.html
	templates/dashboard/splash.html
	templates/projects/gallery.html

+++ apps/feeds/tasks.py
@@ -11,18 +11,34 @@ from django.utils.encoding import smart_str
 from celery.schedules import crontab
 from celery.decorators import periodic_task
 
-from dashboard.models import FeedEntry
+from feeds.models import FeedEntry
 
 log = logging.getLogger(__name__)
 
 
 def parse_entry(entry):
+    """
+    Given a feed entry, return a dictionary with 'title', 'content', 'link',
+    'updated'.
+    """
     title = getattr(entry, 'title', None)
+    if not title:
+        log.debug("Feed entry has no title element.")
+        return None
     content = getattr(entry, 'content', None)
+    # some feed entries have a summary but no content
+    if not content:
+        content = getattr(entry, 'summary', None)
+    if  not content:
+        log.debug("Feed entry has no content or summary element.")
+        return None
     link = getattr(entry, 'link', None)
+    if not link:
+        log.debug("Feed entry has no link element.")
+        return None
     updated = getattr(entry, 'updated_parsed', None)
-    if not (title and content and link and updated):
-        log.debug("No title or no content or no link or no updated")
+    if not updated:
+        log.debug("Feed entry has no updated element.")
         return None
     if type(content) == type([]):
         content = content[0]
@@ -34,26 +50,32 @@ def parse_entry(entry):
     }
 
 
-@periodic_task(run_every=crontab(minute=0, hour=0))
-def update_feeds():
-    feed_url = getattr(settings, 'SPLASH_PAGE_FEED', None)
-    log.debug("update_feeds called with feed url %s" % (feed_url,))
-
+def get_feed_entries(feed_url):
+    """Grab the 4 most recent feed entries from the splash page feed URL."""
+    log.debug('Fetching feed from URL %s' % (feed_url,))
     if not feed_url:
-        log.warn("No feed url defined. Cannot update splash page feed.")
-        return
-
+        log.warn('No feed url defined. Cannot update splash page feed.')
+        return []
     data = urllib2.urlopen(feed_url).read()
     feed = feedparser.parse(data)
     entries = feed.entries[0:4]
+    return entries
 
+
+def parse_feed(feed_url, page):
     ids = []
+    entries = get_feed_entries(feed_url)
+
     for entry in entries:
         parsed = parse_entry(entry)
         if not parsed:
             log.warn("Parsing feed failed. continuing")
             continue
-        body = getattr(parsed['content'], 'value', None)
+        if isinstance(parsed['content'], feedparser.FeedParserDict):
+            if 'value' in parsed['content'].keys():
+                body = parsed['content']['value']
+        else:
+            body = parsed['content']
         if not body:
             log.warn("Parsing feed failed - no body found")
             continue
@@ -66,6 +88,7 @@ def update_feeds():
                     title=parsed['title'].encode('utf-8'),
                     link=parsed['link'].encode('utf-8'),
                     body=cleaned_body,
+                    page=page,
                     checksum=checksum,
                     created_on=time.strftime(
                         "%Y-%m-%d %H:%M:%S", parsed['updated']))
@@ -76,5 +99,19 @@ def update_feeds():
         except:
             log.warn("Encountered an error creating FeedEntry. Skipping.")
             continue
+    return ids
 
-    FeedEntry.objects.exclude(id__in=ids).delete()
+
+@periodic_task(run_every=crontab(minute=0, hour=0))
+def update_feeds():
+    ids = []
+    feeds = getattr(settings, 'FEED_URLS', None)
+    if not feeds:
+        log.debug("No feeds defined, aborting")
+        return
+    for page, feed_url in feeds.iteritems():
+        parsed = parse_feed(feed_url, page)
+        if parsed:
+            ids.extend(parsed)
+    if ids:
+        FeedEntry.objects.exclude(id__in=ids).delete()
